remove old config
